FocusSerial: Implement a `device.reset` command
To make it easier - or in some cases, even possible - to put the keyboard into
programmable mode, we need to be able to initiate a reset from the host side.
For example, ynlike on AVR, where we can send a HUP to the serial port, on GD32,
we need to do it a bit differently: by sending the keyboard a Focus request to
reboot itself.

This implements the command itself.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>