driver::storage::Base: Add an isSliceUninitialized() function
First, we add an `uinitialized_byte` prop to `BaseProps`, so implementations
that don't use `0xff` as default can set it to whatever they use. Do keep in
mind that plenty of code still assumes `0xff` is the uninitialized byte, but
this way we have a starting point.

Then, we add an `isSliceUninitialized()` function, which checks whether a given
slice is completely made up from uninitialized bytes or not.

Fixes #1145.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>