Process Kaleidoscope with `iwyu.py`
This is the result of running the `include-what-you-use` wrapper, followed by
the `clang-format` wrapper on the Kaleidoscope codebase.  It is now safe to use
both without needed any manual corrections after the fact, but it's still
necessary to run clang-format after IWYU, because the two differ in the way they
indent comments after header files.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>