Rename `TEST()` macro to avoid conflicts with gtest
The `TEST()` macro defined in `macro_helpers.h` shared a name with a critical
macro used by gtest code in the simulator, making simulator code very sensitive
to the order of header includes, with rather unhelpful error messages when it
failed.  This change renames the offending macro, and a related one, for good
measure.  Neither renamed macro was directly used by any Kaleidoscope code, so
this doesn't affect any APIs.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>