Fix Macros tap sequence parsing
Both Macros and DynamicMacros were only reading one byte for each `Key` object
in a tap sequence, so it would first read the flags byte of each key in the
sequence and treat it as a keycode byte, using a flags byte of `0`.  As soon as
an unmodified keyboard key was encountered, this would be recognized as the end
of the sequence.  This change fixes the bug by reading and using the flags byte
of each key in the sequence as intended.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>