etc/makefiles/sketch.mk: Copy & symlink .bin files too, if they exist
On GD32, while the build process does create .hex files, we use the .bin ones
for flashing. So copy those to the appropriate place, and do the same symlinking
as for the .elf and .hex files.

Because .bin is not compiled on all platforms, we guard it with an if that
checks for its existence.

There might be a better way to do this, but this was fast and easy.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>