PersistentLEDMode: Add a Focus hook, and a way to disable the auto-save
We'd like to be able to set the default LED mode via Focus, so it can be
configured via Chrysalis. However, we may not want auto-save, so make that
configurable too.

To preserve the EEPROM layout, the highest bit of the previous led mode index
setting was repurposed for the auto save setting. This lets us set the default
mode to anything between 0 and 126 (or 127, if auto save is turned off).

While there, we also add an `onNameQuery` handler, to make it easier for
Chrysalis to detect if the plugin is available.

This addresses the Kaleidoscope parts of keyboardio/Chrysalis#846.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>