PersistentLEDMode: Revert back to auto-saving
This basically reverts 07dcf1dc9be5845ffaab5f409b67f8ccf5e9b5a2, returning the
plugin to its original behaviour of persisting the current led mode. We do this
because we'll be using a new, different plugin to set the default led mode, to
not conflate the two different functionalities.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>