hardware: Fix the bootloader props on a number of devices
A number of devices declared the bootloader in their device properties as
`BootLoader`, while the base class, and anything using it, was looking for
`Bootloader`. This resulted in these devices using the default, dummy
bootloader, rather than the one we intended to set.

This patch corrects that, and everything's using `Bootloader` now, including the
documentation.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>