etc/makefiles: Call bin/focus-send to do the device reset
Instead of simply echoing "device.reset" into the device port, call
`bin/focus-send`, which sets the `raw` setting on the port, before writing into
it. Without that, the command is not recognised by the firmware.

We could do the stty call in the makefile directly, but to support macOS, we'd
need to treat that specially, and at that point, it's easier to just call
`bin/focus-send`. The sketch relies on having the Kaleidoscope repo available
anyway, so we can safely rely on `bin/focus-send` being there, too.

Because we're now using the tool in the makefiles, `focus-test` was renamed to
`focus-send`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>