Focus: correctly handle delayed newline
Previously, a newline that wasn't read with the rest of the command in
a single event would get appended to the command buffer instead of
remaining in the "peek" buffer, so the command dispatch would never
fire.

Stash a space delimiter instead of storing it in the buffer, only to
overwrite it later.

Fix an off-by-one error that could cause a buffered command to not be
null-terminated. This probably didn't cause a problem in practice,
because Focus plugins mostly did a `strcmp` against a fixed string that
is smaller than the command buffer size.

Signed-off-by: Taylor Yu <tlyu@mit.edu>