improve focus-send
Fall back to `dash` on macOS, because `bash` randomly drops serial
input, causing the tool to hang.

Flush the command buffer before sending the requested command. A
failed upload session can cause characters to remain in the command
buffer.

Redirect stdin instead of using a separate file descriptor. Also do
this before running `stty`. This allows the `stty` settings to
actually take effect on macOS, which seems to reset the termios
state of serial devices upon the last close of the device.

Tested on macOS 10.15 and Ubuntu 20.04.

Signed-off-by: Taylor Yu <tlyu@mit.edu>