SpaceCadet: Implement a SpaceCadetConfig plugin
The new plugin enables configuring some aspects of SpaceCadet through Focus: the
current mode, and the global timeout. This is makes it possible to ship firmware
with SpaceCadet included, disabled by default, but still allow one to enable it
without having to map and tap the enable key.

The settings are also persisted into storage.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>