Deprecate FocusLEDCommand
The scope of this plugin is rather wide, which makes it hard for us to change
only some if its behaviour. As in, we'd like the brightness settings to persist,
but we do not want the rest of the things `FocusLEDCommand` does to do so
aswell. It feels wrong to persist only some parts of it.

Thus, deprecate the whole plugin, it will be replaced by smaller ones, with a
more narrow scope, so we can avoid these problems.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>