Because of how makefile rules are evaluated, the previous code was detecting the device port at evaluation time rather than execution time.
This meant that we were doing the "where is the board" check for any
compilation target, even if we'd never flash.

Arduino's board probing is somewhat heavyweight and can take a couple of
seconds.

We move that logic into a shell expression executed at runtime.

On my laptop, this shaves 10 seconds off make -j 9 simulator tests,
which is pretty nice since that used to take about 30 seconds.

But on a plain `make simulator-tests`, it shaved a full minute from the
2 minute and 30 second runtime.