HostPowerManagement: Lift out the suspension check, to remove code duplication
The difference between the AVR and the GD32 implementation was the way we
checked if the device is suspended, otherwise the logic was exactly the same,
duplicated.

To remove the need for duplicating the same code for every architecture we
support, lift out the suspension check instead. This way the core logic becomes
shared between all of them.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>