Correct error in the unused `Layer.deactivateMostRecent()` function
This function was revised to only operate on locked layers, but it's not at all
clear that it should.  Furthermore, it wouldn't necessarily deactivate the top
locked layer; instead, it would deactivate the locked version of a shifted
layer, if that layer was on the stack, leaving the actual top layer unchanged.

This change has it deactivate the top layer, regardless of how it got there.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>