Deprecate unused layer functions
The functions `activateNext()` and `deactivateMostRecent()` are not used by any
core plugins, and don't have much utility now that the layer stack is based on
activation order, and has separate entries for "locked" and "shifted" layers.

`activateNext()` is based on an index-ordered stack, and
`deactivateMostRecent()` could result in confusing behaviour because it's not
clear if it should apply to locked layers, shifted layers, or both.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>