Add alias for Apple "globe" key
This adds an alias for the Apple "globe" key, which triggers the emoji picker on
Apple devices.  It doesn't bear any obvious relationship to the "next keyboard
layout select" terminology in the HID tables, so the alias is warranted.

It is not the same as the Apple keyboard `fn` key, but it is useful nonetheless.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>