MouseKeysConfig: Expose the warp grid size setting
This is a little bit trickier to expose than the others, because the canonical
source of truth is held in `MouseWrapper`, rather than `MouseKeys.settings_`.
While it would be possible to move the property, it's less effort to work around
the situation: `setWarpGridSize()` will simply set both.

Both, because for the sake of convenience, we'll store the property in
`MouseKeys.settings_` too.

To make things safer, `setWarpGridSize()` will sanity check the grid size now,
and won't change it unless the size is supported.

This covers the Kaleidoscope parts of keyboardio/Chrysalis#1174.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>