Re-order plugins for Model100 sketch
The order of plugins in this sketch was somewhat haphazard, and in one case
caused an error (SpaceCadet before Qukeys).  I have reorganized the plugin init
order into categories so that it's easier to tell where plugins are relative to
other plugins that they might interact with.  I've also arranged the plugins
that handle keystrokes in order to minimize potential event-handling problems.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>