Standardize Keyboardio device plugin init order
The Atreus, Model01, and Model100 each had a different order of plugin
initialization in `KALEIDOSCOPE_INIT_PLUGINS()`.  I copied the section from the
Model100 to the other two, and commented out plugins that were not in use in the
default sketches.  I also removed the LED plugin section from the Atreus sketch,
since it has no LEDs.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>