Focus: Add a mechanism to discover commands eligible for backup
The goal here is to make it easier for Chrysalis and other tools to make
structured backups. Chrysalis currently has a hard-coded list of commands to run
and backup the results of - not ideal. We want a solution where a tool can
automatically discover what can be backed up safely.

We can't just call every command in the `help` output without an argument,
because some of them (`device.reset`, `eeprom.erase`, etc) are destructive. We
do not wish to maintain a deny list, either.

We'd like a mechanism where plugins themselves can tell which commands they can
offer for backing up. This patch implements such a solution, one very similar to
how we handle `help`. In addition to responding to `help`, plugins can now
respond to `backup`, and write a list of commands in the same format as for
`help`: one per line.

The tool on the host side can take this list of commands, execute them, and
store their values for backup purposes. Simple, straightforward, lightweight.

On top of adding the mechanism, the patch also changes all bundled plugins to
respond to the `backup` command if appropriate, with the commands that should be
backed up.

Fixes #1279.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>