IdleLEDs: Set the default limit in PersistentIdleLEDs correctly
When the EEPROM is empty, and we're setting the default idle time limit, we need
to set that in seconds, rather than millis, so divide the internal time
limit (uint32_t, in millis) by 1000 so it fits into the uint16_t we use for
persisting the time limit in seconds.

This correctly sets the default to 10 minutes, rather than a very awkward 2
hours, 49 minutes,  36 seconds (10176 seconds).

Signed-off-by: Gergely Nagy <algernon@keyboard.io>