IdleLEDs: Implement a .suspend() and a .resume() method
Sometimes we want to control LEDs outside of `IdleLEDs`, like through
`HostPowerManagement`, and in these cases, we would like `IdleLEDs` to have an
up-to-date idea about what's going on. That is, if the host enters sleep,
`IdleLEDs` should be aware of that, and enter idle state. Once the host wakes
up, either via the keyboard or otherwise, `IdleLEDs` should resume its tasks.

With the new `.suspend()` and `.resume()` method, this becomes possible.

This addresses a big part of #1287.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>