Atreus: Handle SpaceCadetConfig properly
Rather than disabling `SpaceCadet` in `setup()`, let the `SpaceCadetConfig`
plugin handle the initial state, unless it is in an unconfigured state, in which
case we should default to disabling `SpaceCadet`.

This mirrors the logic we do on the Model100, and makes the persisted
configuration actually useful.

Reported by esclee92 on Discord, thank you!

Signed-off-by: Gergely Nagy <algernon@keyboard.io>