fix Model 100 flashing on Linux
Flashing on Model 100 using `make flash` would fail with a cryptic
message like
```
Error during Upload: main file missing from sketch
```
if the keyboard was already in bootloader mode.

The `$(shell)` function in `sketch.mk` that parsed the board listing
for constructing the upload command line would produce `--port` without
an argument, which meant that the `--input-dir` would get interpreted
as the argument to `--port`. This was due to an OS-related behavior
difference of `xargs` when given empty input.

The reason this wasn't happening when the keyboard was in application
mode was that `make` expanded the `$(shell)` function before beginning
to execute the recipe, so it would successfully get the port before the
keyboard rebooted into bootloader mode.

Thanks to @dakkar and @Lilith-Daemon for help with troubleshooting and
testing solutions.

Signed-off-by: Taylor Yu <code@argon.blue>