AutoShiftConfig: Add a .disableAutoShiftIfUnconfigured() method
The AutoShift plugin is enabled by default, but when using AutoShiftConfig, we
may want to have it disabled by default, so that we can include it in firmware
updates without turning optional, behaviour changing features on by default. The
new method helps us accomplish this, in a way similar to how SpaceCadet
accomplishes the same thing.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>