fallback keymap
Make `EEPROMSettings` fall back to a usable key map if it detects
a possibly incompatible EEPROM layout. This forces
`keymap.onlyCustom 0` and `settings.defaultLayer 0`, but does
not write them to EEPROM.

A new `settings.acceptInvalid` Focus command may be used to
accept the current settings as valid, updating the checksum.

Signed-off-by: Taylor Yu <code@argon.blue>