make fit on Atreus
Adapt previous to fit on the Atreus example sketch. Notably, this
makes `EEPROMKeymap::max_layers()` private so that it can be inlined.
Therefore, sketches like the `EEPROM-Keymap-Programmer` example sketch
that used that method needed to be adjusted. That method seemed obsolete
anyway, because it doesn't do some error checking that
`EEPROMKeymap::setup()` does.

Also, it removes the `settings.acceptInvalid` command, replacing it
with an optional argument to `settings.valid?`. Unfortunately, this
means that Chrysalis has no direct way to detect the presence of this
capability.

Signed-off-by: Taylor Yu <code@argon.blue>