Further separate layer shifts and locks
When a layer lock key is pressed, ignore shifted layers in the stack when
determining if the target layer is on the top of the layer stack.  This lets
users put a `LockLayer(N)` on a layer that is accessed by holding
`ShiftToLayer(M)`, and have more consistent behaviour.  Previously, toggling
layer `N` on worked as expected, but it couldn't be toggled off with a single
tap because the shifted layer `M` would be on top when the `LockLayer(N)` was
tapped.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>