reset HID protocol on USB reset
Reset the HID protocol on USB reset, as required by the HID
specification.

This approach adds polling methods in the MCU driver. The default
implementation checks for changes in the USB configuration state,
because that's the most portable at the moment. Future work can use USB
reset hooks, if available.

Signed-off-by: Taylor Yu <code@argon.blue>