fix redundant setup code on Model 100
Remove some redundant calls in `Model100::setup()` that caused
`Model100KeyScanner::enableScannerPower()` to be called three
times during startup. Reduce the startup delay to 70ms to allow
the ATtinys to boot.

For some reason, the delay is still needed, because the I2C seems
to hang, otherwise. The cause of the hang is still under
investigation.

Signed-off-by: Taylor Yu <code@argon.blue>