Fix defining multiple prefix layers
The `setPrefixLayers()` function had a bug where it was treating a pointer as an
array with a known size, which was incorrect.  Instead of passing a pointer to
the array as an argument, we now pass a reference to the array, as an array type
with a known fixed size by using a template function.  Now it's possible to have
multiple prefix layers instead of just one.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>