Add testcase for Issue #1324
This testcase verifies that multiple prefix layers can be defined.  We need all
three prefix layers in order to demonstrate the bug because of the different
sizes of a pointer in the simulator and on an AVR processor (8 vs 2).  On
physical hardware, I expect this to fail if more than one prefix layer is
defined.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>