Make autoshift longpress configurable
This commit allows specifying the keys to produce on long press if not a
shifted, but a totally different character is wanted.

Such an explicit mapping takes precedence over shifting the key. That
means if all alpanumeric characters are configured for AutoShift, but
the ‘e’ key has an explicit mapping to produce ‘ë’, a long press on ‘e’
will result in ’ë’, not ‘E’.

Signed-off-by: Marco Herrn <marco@mherrn.de>