use pointer collection in mouse descriptors
It's possible that some KVM switches or OSes didn't like the mouse
devices, because the axes aren't wrapped in a Pointer Physical
Collection.

Also use the correct macro for the item tag for the multibyte AC Pan on
the relative mouse. No change in encoding; it's just a more meaningful
macro.

Signed-off-by: Taylor Yu <code@argon.blue>