WS2812 Neopixel driver - workaround parasitic current of 0.5mA on nRF52 when digital io pin is left driven low.
LED WS2812 undo early output pin disconnect - only flop our putput pin to disconnected as we go to sleep. This burns power while we are awake, but on MP keyboardio Preonic keyboards it leads to bright flashing lights for an as-yet-undiagnosable reason.
This change needs additional support in the device's power management routines

Signed-off-by: Jesse Vincent <jesse@fsck.com>